Changes to add Poisson fitting.
